+++ tests/e2e/game/full-game.test.js
@@ -90,6 +90,8 @@ describe("B - Full game of 31 players with all roles", () => {
                 expect(game.options.roles.areRevealedOnDeath).to.be.true;
                 expect(game.options.roles.sheriff.electedAt.turn).to.be.equal(1);
                 expect(game.options.roles.sheriff.electedAt.phase).to.be.equal("night");
+                expect(game.options.roles.sheriff.hasDoubledVote).to.be.true;
+                expect(game.options.roles.sheriff.canSettleVotes).to.be.true;
                 expect(game.options.roles.bigBadWolf.isPowerlessIfWerewolfDies).to.be.true;
                 expect(game.options.roles.whiteWerewolf.wakingUpInterval).to.be.equal(2);
                 expect(game.options.roles.seer.canSeeRoles).to.be.true;
@@ -101,6 +103,8 @@ describe("B - Full game of 31 players with all roles", () => {
                 expect(game.options.roles.stutteringJudge.voteRequestsCount).to.equal(1);
                 expect(game.options.roles.wildChild.isTransformationRevealed).to.be.false;
                 expect(game.options.roles.dogWolf.isChosenSideRevealed).to.be.false;
+                expect(game.options.roles.dogWolf.isChosenSideRandom).to.be.false;
+                expect(game.options.roles.thief.isChosenCardRevealed).to.be.false;
                 expect(game.options.roles.thief.mustChooseBetweenWerewolves).to.be.true;
                 expect(game.options.roles.thief.additionalCardsCount).to.be.equal(2);
                 expect(game.options.roles.piedPiper.charmedPeopleCountPerNight).to.equal(2);